requirements.txt
click
fastapi[all]
hydra-core
joblib
motor
omegaconf
pdf2image
Pillow
pymongo
PyMuPDF
python-multipart
urllib3
pandas
boto3
numpy
albumentations
efficientnet-pytorch
opencv-python
